+++ sdk/storage/storage-queue/src/utils/constants.ts
@@ -1,8 +1,8 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-export const SDK_VERSION: string = "12.25.0";
-export const SERVICE_VERSION: string = "2025-01-05";
+export const SDK_VERSION: string = "12.26.0";
+export const SERVICE_VERSION: string = "2025-05-05";
 
 /**
  * The OAuth scope to use with Azure Storage.