+++ sdk/ai/ai-inference-rest/CHANGELOG.md
@@ -1,11 +1,25 @@
 # Release History
 
+## 1.0.0-beta.6 (2025-03-18)
+
+### Features Added
+
+- Add "developer" chat role, and ChatRequestDeveloperMessage interface.
+- Added audio input (data and url) to chat completions, with samples.
+
+
+### Bugs Fixed
+
+* Re-export ErrorModel type from azure-rest/core-client (#33333)
+
 ## 1.0.0-beta.5 (2025-01-22)
 
 ### Breaking Changes
+
 - Change ChatCompletionsResponseFormatJSON class name to ChatCompletionsResponseFormatJsonObject.
 
 ### Features Added
+
 - Add samples and tests for image embeddings.
 - Add ChatCompletionsResponseFormatJsonSchema class for structured output in completions response.
 - Add sample for Chat Completions with structured output
@@ -19,14 +33,17 @@
 ## 1.0.0-beta.3 (2024-10-22)
 
 ### Features Added
+
 - Adds support for instrumentation using OpenTelemetry for chat completion without streaming.
 
 ## 1.0.0-beta.2 (2024-07-17)
 
 ### Features Added
+
 - Adds support for text and image embeddings.
 
 ## 1.0.0-beta.1 (2024-06-30)
 
 ### Features Added
+
 - This is the initial beta release for the Azure AI Inference SDK, which is a common API supporting Azure AI Inference models.